Upgrade Go dependencies to latest versions
This commit involves upgrading various Go dependencies to their latest versions. The dependencies include 'github.com/Azure', 'github.com/cpuguy83', 'golang.org' packages and others. The upgrades are necessary to ensure the code uses the latest features and bug fixes available in these dependencies. This is crucial for maintaining the overall quality and agility of the software.